+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - OAS 3.0 is not available (issue #58)
 - Incorrect schema version in the metadata record (issue #59) 
+- XML validation is not threadsafe. (issue #60)
 
 ## [0.3.4] - 2021-11-17
 ### Fixed